Rename package to data_service #2714
