+++ data-serving/reusable-data-service/README.rst
@@ -6,6 +6,6 @@ Introduction
 ============
 
 The Global.health reusable data service takes lessons learned from the
-COVID-19 data service (see ../data-service) and provides a scalable,
+COVID-19 data service (see ../data_service) and provides a scalable,
 extensible application server for CRUD operations on line list data
 for infectious disease outbreaks.
